Close every showstopper from the fresh-eyes audit
A second clean-room agent walked the workflow from scratch and surfaced
four real bugs that would frustrate a real student. All four are now
fixed, plus a related UX polish.

Showstoppers fixed:

1. ``generate_control_values.py --verify --src tutorials --chapters model``
   used to ModuleNotFoundError because prepare_tutorials.py never copied
   ``solutions/model/control_values/`` to ``tutorials/``. Add the dir to
   FOLDER_COPY_PATHS.

2. README quickstart cited ``jupyter notebook solutions/overview.ipynb``
   but the file lives at ``solutions/model/overview.ipynb``. Fixed in
   both README.md and README.en.md.

3. ``_generate.py`` chained ``.double()`` onto module constructors, so
   any unfilled ``__init__`` (still ``pass``) blew up with a raw 17-line
   PyTorch AttributeError instead of the project's friendly hint.
     - Strip every ``.double()`` chained on a module constructor from
       all six chapter ``_generate.py`` files (the ``controlled_execution``
       helper doubles internally; the chained call was redundant).
     - Add ``_detect_uninitialised`` in ``runtime/checks.py`` and call
       it from both ``test_module_shape`` and ``controlled_execution``
       so the friendly message fires on either code path.
     - Wrap ``build_tiny_protenix()`` in ``model/control_values/_generate.py``
       so even sub-module __init__ failures during full Protenix
       construction surface as a clear "fill in the per-chapter blanks
       first" message.

4. ``check_solutions.py`` defaulted to fail-fast, contradicting the
   README claim that it "runs every chapter in one shot". Flip the
   default to keep-going + summary, add ``--fail-fast`` for CI.

Friction polished:

- attention.ipynb section 1.1 used to bundle Linear / LinearNoBias /
  BiasInitLinear into one cell with a single ``print('Linear ✓')``.
  Split into 1.1 (Linear), 1.2 (LinearNoBias), 1.3 (BiasInitLinear) so
  students get incremental green checkmarks. Renumber 1.4–1.8 below.

- ``lessons/README.md`` updated to clearly say "WIP — TODO pseudocode in
  solutions/ is the primary reference until each writeup lands"; both
  README.md and README.en.md mark the lessons row of the trio table
  with **(WIP)** so newcomers don't expect content there.

Verified:

  python check_solutions.py                   # solutions/ 6/6 PASS
  python generate_control_values.py --verify  # 6/6 PASS
  python -m model.inference …                 # pLDDT 33.82 unchanged

  python check_solutions.py --src tutorials   # processes all 6 chapters
                                              # before summary; feature_extraction
                                              # passes (read-only), the 5 lab
                                              # chapters FAIL with friendly errors

  python generate_control_values.py --verify --src tutorials --chapters model
      # now also fails cleanly with a friendly "fill in __init__ first"

+++ solutions/attention/control_values/_generate.py
@@ -49,7 +49,7 @@ def main(overwrite: bool = True) -> None:
     torch.manual_seed(0)
 
     # ----- Linear (initializer="default") ----------------------------------
-    lin = Linear(in_features=c_a, out_features=c_z).double()
+    lin = Linear(in_features=c_a, out_features=c_z)
     test_module_shape(lin, "linear_default", CONTROL_FOLDER, overwrite_results=overwrite)
     test_module_forward(
         lin, "linear_default",
@@ -60,7 +60,7 @@ def main(overwrite: bool = True) -> None:
     )
 
     # ----- LinearNoBias ----------------------------------------------------
-    lin_nb = LinearNoBias(in_features=c_a, out_features=c_z).double()
+    lin_nb = LinearNoBias(in_features=c_a, out_features=c_z)
     test_module_shape(lin_nb, "linear_nobias", CONTROL_FOLDER, overwrite_results=overwrite)
     test_module_forward(
         lin_nb, "linear_nobias",
@@ -71,7 +71,7 @@ def main(overwrite: bool = True) -> None:
     )
 
     # ----- BiasInitLinear (biasinit=-2.0) ----------------------------------
-    bil = BiasInitLinear(in_features=c_s, out_features=c_a, bias=True, biasinit=-2.0).double()
+    bil = BiasInitLinear(in_features=c_s, out_features=c_a, bias=True, biasinit=-2.0)
     test_module_shape(bil, "bias_init_linear", CONTROL_FOLDER, overwrite_results=overwrite)
     test_module_forward(
         bil, "bias_init_linear",
@@ -82,7 +82,7 @@ def main(overwrite: bool = True) -> None:
     )
 
     # ----- LayerNorm (default scale + offset) ------------------------------
-    ln = LayerNorm(c_a).double()
+    ln = LayerNorm(c_a)
     test_module_shape(ln, "layer_norm", CONTROL_FOLDER, overwrite_results=overwrite)
     test_module_forward(
         ln, "layer_norm",
@@ -115,7 +115,7 @@ def main(overwrite: bool = True) -> None:
         gating=True, q_linear_bias=True,
         use_efficient_implementation=False,
         zero_init=False,
-    ).double()
+    )
     test_module_shape(attn, "mha_gated", CONTROL_FOLDER, overwrite_results=overwrite)
     test_module_method(
         attn, "mha_gated",
@@ -127,7 +127,7 @@ def main(overwrite: bool = True) -> None:
     )
 
     # ----- AdaptiveLayerNorm ----------------------------------------------
-    adaln = AdaptiveLayerNorm(c_a=c_a, c_s=c_s).double()
+    adaln = AdaptiveLayerNorm(c_a=c_a, c_s=c_s)
     test_module_shape(adaln, "adaptive_layer_norm", CONTROL_FOLDER, overwrite_results=overwrite)
     test_module_forward(
         adaln, "adaptive_layer_norm",
@@ -138,7 +138,7 @@ def main(overwrite: bool = True) -> None:
     )
 
     # ----- Transition ------------------------------------------------------
-    tr = Transition(c_in=c_a, n=n_factor).double()
+    tr = Transition(c_in=c_a, n=n_factor)
     test_module_shape(tr, "transition", CONTROL_FOLDER, overwrite_results=overwrite)
     test_module_forward(
         tr, "transition",
@@ -153,7 +153,7 @@ def main(overwrite: bool = True) -> None:
         has_s=True, create_offset_ln_z=False,
         n_heads=N_head, c_a=c_a, c_s=c_s, c_z=c_z,
         biasinit=-2.0, cross_attention_mode=False,
-    ).double()
+    )
     # F.scaled_dot_product_attention rejects mixed dtypes; the explicit path
     # is dtype-tolerant. We turn the efficient path off only for testing.
     # F.scaled_dot_product_attention 不容忍 dtype 不一致，
